refactor: move token, aws-S3 to models/
